Updates to run on Fedora.

+++ tasks/torpedo.rake
@@ -15,11 +15,22 @@ task :torpedo do
 	out=%x{
 ssh #{SSH_OPTS} root@#{gw_ip} bash <<-"BASH_EOF"
 ssh #{server_name} bash <<-"EOF_SERVER_NAME"
+
+if [ -f /bin/rpm ]; then
+	rpm -q rubygems &> /dev/null || yum install -y rubygems &> /dev/null
+	rpm -q rubygem-json &> /dev/null || yum install -y rubygem-json &> /dev/null
+fi
 if ! gem list | grep torpedo &> /dev/null; then
 	gem install --no-rdoc --no-ri torpedo
-	ln -sf /var/lib/gems/1.8/gems/torpedo-*/bin/torpedo /usr/bin/torpedo
+	# Ubuntu fails to link it into /bin
+	[ ! -f /usr/bin/torpedo ] && ln -sf /var/lib/gems/1.8/gems/torpedo-*/bin/torpedo /usr/bin/torpedo
 fi
-[ -f /root/novarc ] && source /root/novarc
+if [ ! -f /root/novarc ]; then
+cd /root
+nova-manage project zipfile nova admin
+unzip -o nova.zip
+fi
+source /root/novarc
 if [ -f /root/openstackrc ]; then
 	if ! grep EC2_SECRET_KEY /root/openstackrc &> /dev/null; then
 		echo "export EC2_SECRET_KEY=\"admin\"" >> /root/openstackrc
@@ -36,7 +47,11 @@ if [[ "#{mode}" == "libvirt" ]]; then
 	# When using libvirt we use an AMI style image which require keypairs
 	export KEYPAIR="/root/test.pem"
 	export KEYNAME="test"
-	dpkg -l euca2ools &> /dev/null || apt-get install -q -y euca2ools &> /dev/null
+	if [ -f /bin/rpm ]; then
+		rpm -q euca2ools &> /dev/null || yum install -y euca2ools &> /dev/null
+	else
+		dpkg -l euca2ools &> /dev/null || apt-get install -q -y euca2ools &> /dev/null
+	fi
 	[ -f "$KEYPAIR" ] || euca-add-keypair "$KEYNAME" > "$KEYPAIR"
 	chmod 600 /root/test.pem
 	cat > ~/.torpedo.conf <<-EOF_CAT
